# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "ConversationsDeletedOut": {
        "properties": {
          "deleted": {
            "title": "Deleted",
            "type": "integer"
          },
          "kept": {
            "title": "Kept",
            "type": "integer"
          }
        },
        "required": [
          "deleted",
          "kept"
        ],
        "title": "ConversationsDeletedOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MeOut": {
        "description": "The current user with their membership \u2014 the /orgs/me payload.",
        "properties": {
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "full_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Full Name"
          },
          "org": {
            "$ref": "#/components/schemas/OrgOut"
          },
          "role": {
            "enum": [
              "owner",
              "clinician",
              "viewer"
            ],
            "title": "Role",
            "type": "string"
          },
          "specialty": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Specialty"
          },
          "user_id": {
            "format": "uuid",
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "user_id",
          "email",
          "full_name",
          "specialty",
          "role",
          "org"
        ],
        "title": "MeOut",
        "type": "object"
      },
...
      "OrgOut": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "plan": {
            "enum": [
              "free",
              "pro",
              "enterprise"
            ],
            "title": "Plan",
            "type": "string"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "slug",
          "plan",
          "created_at"
        ],
        "title": "OrgOut",
        "type": "object"
      },
...
      "PreferencesOut": {
        "description": "What the assistant, voice and Learn do by default for this person.\n\n``saved`` is False until the person first changes something: until then\nthese are the product's defaults.",
        "properties": {
          "audience": {
            "default": "patient",
            "enum": [
              "patient",
              "clinician",
              "student"
            ],
            "title": "Audience",
            "type": "string"
          },
          "followed_specialties": {
            "items": {
              "type": "string"
            },
            "title": "Followed Specialties",
            "type": "array"
          },
          "learn_depth": {
            "default": "auto",
            "enum": [
              "auto",
              "mbbs",
              "pg"
            ],
            "title": "Learn Depth",
            "type": "string"
          },
          "learn_scope": {
            "default": "all",
            "enum": [
              "all",
              "mbbs",
              "pg"
            ],
            "title": "Learn Scope",
            "type": "string"
          },
          "saved": {
            "default": false,
            "title": "Saved",
            "type": "boolean"
          },
          "show_timeline": {
            "default": true,
            "title": "Show Timeline",
            "type": "boolean"
          },
          "sources_open": {
            "default": false,
            "title": "Sources Open",
            "type": "boolean"
          },
          "updated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "voice_continuous": {
            "default": true,
            "title": "Voice Continuous",
            "type": "boolean"
          },
          "voice_name": {
            "default": "thalia",
            "title": "Voice Name",
            "type": "string"
          },
          "voice_rate": {
            "default": 1.0,
            "title": "Voice Rate",
            "type": "number"
          }
        },
        "title": "PreferencesOut",
        "type": "object"
      },
      "PreferencesUpdate": {
        "additionalProperties": false,
        "description": "Only the fields sent change; send ``[]`` to clear followed specialties.",
        "properties": {
          "audience": {
            "anyOf": [
              {
                "enum": [
                  "patient",
                  "clinician",
                  "student"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Audience"
          },
          "followed_specialties": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "maxItems": 60,
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Followed Specialties"
          },
          "learn_depth": {
            "anyOf": [
              {
                "enum": [
                  "auto",
                  "mbbs",
                  "pg"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Learn Depth"
          },
          "learn_scope": {
            "anyOf": [
              {
                "enum": [
                  "all",
                  "mbbs",
                  "pg"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Learn Scope"
          },
          "show_timeline": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Show Timeline"
          },
          "sources_open": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sources Open"
          },
          "voice_continuous": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Voice Continuous"
          },
          "voice_name": {
            "anyOf": [
              {
                "maxLength": 20,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Voice Name"
          },
          "voice_rate": {
            "anyOf": [
              {
                "maximum": 1.5,
                "minimum": 0.75,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Voice Rate"
          }
        },
        "title": "PreferencesUpdate",
        "type": "object"
      },
...
      "ProfileUpdate": {
        "additionalProperties": false,
        "description": "A person's own name and specialty. An empty string clears the field.",
        "properties": {
          "full_name": {
            "anyOf": [
              {
                "maxLength": 120,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Full Name"
          },
          "specialty": {
            "anyOf": [
              {
                "maxLength": 120,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Specialty"
          }
        },
        "title": "ProfileUpdate",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Evidence-grounded clinical question answering over public medical literature.",
    "title": "ClinicalContext AI",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/me/conversations": {
      "delete": {
        "description": "Delete all your assistant conversations (Chat, Learn, Treatment, voice).\n\nEvidence searches stay: they are your organization's record. So does any\nconversation with an answer saved to a binder, shared publicly, or with\nrecorded versions; ``kept`` counts them.",
        "operationId": "delete_my_conversations_api_v1_me_conversations_delete",
        "parameters": [
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversationsDeletedOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete My Conversations",
        "tags": [
          "settings"
        ]
      }
    },
    "/api/v1/me/export": {
      "get": {
        "description": "A JSON copy of your profile, preferences, conversations and feedback.",
        "operationId": "export_my_data_api_v1_me_export_get",
        "parameters": [
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Export My Data",
        "tags": [
          "settings"
        ]
      }
    },
    "/api/v1/me/preferences": {
      "get": {
        "description": "This person's defaults for the assistant, voice and Learn.",
        "operationId": "get_preferences_api_v1_me_preferences_get",
        "parameters": [
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreferencesOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Preferences",
        "tags": [
          "settings"
        ]
      },
      "patch": {
        "description": "Change some preferences; the fields not sent keep their values.",
        "operationId": "update_preferences_api_v1_me_preferences_patch",
        "parameters": [
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreferencesUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreferencesOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Preferences",
        "tags": [
          "settings"
        ]
      }
    },
    "/api/v1/me/profile": {
      "patch": {
        "description": "Change your name or specialty. Role and organization are not yours to set.",
        "operationId": "update_profile_api_v1_me_profile_patch",
        "parameters": [
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Profile",
        "tags": [
          "settings"
        ]
      }
    },
...
  }
}